// src/app/core/types/ScorecardDetails.ts
export interface ScorecardDetailsIf {
    cfType: string;
    cfStat: string;
    ownership: string;
    crcy: string;
    amountGranted: number;
    currentBalance: number;
    arrearsAmount: number;
    installmentAmount: number;
    amountWrittenOff: number;
    lastUpdatedDate: string;
    firstDisburseDate: string;
    latestPaymentDate: string;
    contractSecurityType: string;
    entityCode: string;
}